==> code/03-Analyse.Rmd <==
---
title: "Analyse"
author: "JJayes"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---


For google patents

Map of where patents are registered. 

```{r}
theme_set(theme_light())

df <- read_rds(here("data", "patent-info-google", "Electricity_patents_list_Google_Patents.rds"))
```

```{r}
df_google_counts <- df %>% 
  mutate(year = lubridate::year(priority_date)) %>% 
  count(year) %>% 
  mutate(source = "Google")

df_google_counts %>% 
  ggplot(aes(year, n)) +
  geom_col(fill = "midnightblue") +
  scale_y_continuous(labels = scales::number_format()) +
  labs(y = "Patents including the term 'Electricity'",
       x = NULL)
```


```{r}
df_1 <- read_rds("data/patent-info-prv/Electricity_patents_list_prv_19-01-22.rds")

df_prv_counts <- df_1 %>% 
  mutate(year = lubridate::year(filing_date)) %>% 
  count(year) %>% 
  mutate(source = "Patent- och registreringsverket")

df_prv_counts %>% 
  ggplot(aes(year, n)) +
  geom_col(fill = "midnightblue") +
  scale_y_continuous(labels = scales::number_format()) +
  labs(y = "Patents including the term 'Electricity'",
       x = NULL)
```


```{r}
jpeg(
  filename = here::here("figures", "patent_counts_by_source.jpeg"),
  height = 4,
  width = 6,
  units = "in",
  res = 1000
)

df_prv_counts %>% 
  bind_rows(df_google_counts) %>% 
  ggplot(aes(year, n, fill = source)) +
  geom_col() +
  scale_fill_brewer(palette = "Paired") +
  facet_wrap(~ source, nrow = 2) +
  theme(legend.position = "none") +
  labs(y = "Patents including the term 'Electricity'",
       x = NULL)

dev.off()
```


```{r}
library(tidytext)

df_1 %>% 
  unnest_tokens(title, output = "word") %>% 
  mutate(decade = lubridate::year(filing_date) - lubridate::year(filing_date) %% 10) %>% 
  count(word, decade, sort = T)
  # %>% write_excel_csv("data/patent-info-prv/prv_words.csv")

df_1_words <- read_excel("data/patent-info-prv/prv_words_translated.xlsx")

df_1_words <- df_1_words %>% 
  rename(sw_word = word...1, eng_word = word...4)
```


```{r}
jpeg(
  filename = here::here("figures", "patent_title_words_prv.jpeg"),
  height = 4,
  width = 6,
  units = "in",
  res = 1000
)


df_1_words %>% 
  filter(decade != 2020) %>% 
  anti_join(stop_words, by = c("eng_word" = "word")) %>% 
  filter(eng_word != "NA") %>% 
  group_by(decade) %>% 
  slice_max(n, n = 12) %>% 
  ungroup() %>% 
  mutate(eng_word = reorder_within(eng_word, n, decade)) %>% 
  ggplot(aes(n, eng_word, fill = factor(decade))) +
  geom_col(show.legend = F) +
  scale_y_reordered() +
  scale_fill_viridis_d() +
  facet_wrap(~ decade, scales = "free_y") +
  labs(x = "Number of times term is used in patent title",
       y = NULL)

dev.off()
```

```{r}
df_1 %>% 
  filter(!is.na(applicant)) %>% 
  mutate(applicant = str_to_title(applicant)) %>% 
  count(applicant, sort = T) %>% 
  head(10) %>% 
  gt() %>% 
  cols_label(applicant = md("**PRV Patent applicants**"),
             n = md("**Number of applications**, 1880-2021")) 
```



### More from google patents with the scraped data

```{r}
df <- read_rds("data/patent-info-google/patent_df_cleaned_1.rds") %>% 
  bind_rows(read_rds("data/patent-info-google/patent_df_cleaned_2.rds"))
```


```{r}
df %>% 
  select(countries) %>% 
  mutate(n = str_count(countries, ",")) %>% 
  count(n) %>% 
  tail

df_countries <- df %>% 
  select(countries) %>% 
  separate(countries, into = c(paste0("ctry_", rep(1:45, 1))))
```

```{r}
library(recipes)

dummy_multi_choice_rec <- recipe(~ ., data = df_countries) %>%
  step_dummy_multi_choice(starts_with("ctry_")) %>%
  prep()

df_countries <- bake(dummy_multi_choice_rec, new_data = NULL)
```

```{r}
library(corrr)
df_countries %>%
  select(starts_with("ctry_")) %>%
  rename_with(.cols = everything(), ~ str_remove(.x, "ctry_1_")) %>%
  correlate() %>% 
  filter(term != "SE") %>% 
  select(-SE) %>% 
  rearrange() %>%
  shave() %>%
  rplot(shape = 15, colours = c("midnightblue", "white", "darkorange"))

# df %>%
#   select(starts_with("ing_1")) %>%
#   rename_with(.cols = everything(), ~ str_remove(.x, "ing_1_")) %>% 
#   correlate() %>% 
#   network_plot()
```

Maybe filter it more

```{r}
df_countries <- df %>% 
  select(countries)

countries_12 <- df_countries %>% 
  separate_rows(countries) %>% 
  count(countries) %>% 
  slice_max(n, n = 13) %>% 
  filter(countries != "SE") %>% 
  pull(countries)

df_countries <- df_countries %>% 
  mutate(ctry_de = str_detect(countries, "DE"),
         ctry_us = str_detect(countries, "US"),
         ctry_gb = str_detect(countries, "GB"),
         ctry_fr = str_detect(countries, "FR"),
         ctry_jp = str_detect(countries, "JP"),
         ctry_it = str_detect(countries, "IT"),
         ctry_ca = str_detect(countries, "CA"),
         ctry_nl = str_detect(countries, "NL"),
         ctry_ch = str_detect(countries, "CH"),
         ctry_be = str_detect(countries, "BE"),
         ctry_es = str_detect(countries, "ES"),
         ctry_au = str_detect(countries, "AU"))
```



```{r}
library(corrr)

jpeg(
    filename="figures/patent_country_corrs_top_12.jpeg",
    width=6,
    height=4,
    units="in",
    res=1000)

df_countries %>%
  select(starts_with("ctry_")) %>%
  rename_with(.cols = everything(), ~ str_remove(.x, "ctry_")) %>%
  rename_with(., toupper) %>% 
  correlate() %>% 
  rearrange() %>%
  shave() %>%
  rplot(shape = 15, colours = c("white", "white", "midnightblue"), print_cor = T)

dev.off()
```


```{r}
df_countries %>%
  select(starts_with("ctry_")) %>%
  rename_with(.cols = everything(), ~ str_remove(.x, "ctry_")) %>% 
  correlate() %>% 
  network_plot(colours = c("blue", "orange", "darkred"))
```
